+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,6 +1,7 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.techcourse;
 
 import com.interface21.web.WebApplicationInitializer;
+import com.interface21.webmvc.servlet.mvc.tobe.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletRegistration;
 import org.slf4j.Logger;
