+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/DispatcherServlet.java
@@ -25,7 +25,7 @@ public class DispatcherServlet extends HttpServlet {
 
     @Override
     public void init() {
-        handlerMappingRegistry = new HandlerMappingRegistry(DispatcherServlet.class.getPackage());
+        handlerMappingRegistry = new HandlerMappingRegistry();
         handlerAdapterRegistry = new HandlerAdapterRegistry();
     }
